move presto schema sql into impl
